Create Polygon-API folder and move API endpoints index file and date functions to utils file
